requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.6
cron_descriptor==2.0.6
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==37.8.0
flower==2.0.1
freezegun==1.5.5
humanize==4.13.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
MarkupSafe==3.0.2
model-bakery==1.20.5
packaging==25.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
